+++ broker/src/main/kotlin/stores/dbs/sqlite/MetricsStoreSQLite.kt
@@ -201,21 +201,21 @@ class MetricsStoreSQLite(
                 val row = results.getJsonObject(0)
                 val metricsJson = JsonObject(row.getString("metrics"))
                 BrokerMetrics(
-                    messagesIn = metricsJson.getLong("messagesIn", 0L),
-                    messagesOut = metricsJson.getLong("messagesOut", 0L),
+                    messagesIn = metricsJson.getDouble("messagesIn", 0.0),
+                    messagesOut = metricsJson.getDouble("messagesOut", 0.0),
                     nodeSessionCount = metricsJson.getInteger("nodeSessionCount", 0),
                     clusterSessionCount = metricsJson.getInteger("clusterSessionCount", 0),
                     queuedMessagesCount = metricsJson.getLong("queuedMessagesCount", 0L),
                     topicIndexSize = metricsJson.getInteger("topicIndexSize", 0),
                     clientNodeMappingSize = metricsJson.getInteger("clientNodeMappingSize", 0),
                     topicNodeMappingSize = metricsJson.getInteger("topicNodeMappingSize", 0),
-                    messageBusIn = metricsJson.getLong("messageBusIn", 0L),
-                    messageBusOut = metricsJson.getLong("messageBusOut", 0L),
+                    messageBusIn = metricsJson.getDouble("messageBusIn", 0.0),
+                    messageBusOut = metricsJson.getDouble("messageBusOut", 0.0),
                     timestamp = metricsJson.getString("timestamp") ?: at.rocworks.extensions.graphql.TimestampConverter.currentTimeIsoString()
                 )
             } else {
                 // No historical data found, return zero metrics
-                BrokerMetrics(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, at.rocworks.extensions.graphql.TimestampConverter.currentTimeIsoString())
+                BrokerMetrics(0.0, 0.0, 0, 0, 0, 0, 0, 0, 0.0, 0.0, at.rocworks.extensions.graphql.TimestampConverter.currentTimeIsoString())
             }
         })
     }
@@ -265,13 +265,13 @@ class MetricsStoreSQLite(
                 val row = results.getJsonObject(0)
                 val metricsJson = JsonObject(row.getString("metrics"))
                 SessionMetrics(
-                    messagesIn = metricsJson.getLong("messagesIn", 0L),
-                    messagesOut = metricsJson.getLong("messagesOut", 0L),
+                    messagesIn = metricsJson.getDouble("messagesIn", 0.0),
+                    messagesOut = metricsJson.getDouble("messagesOut", 0.0),
                     timestamp = metricsJson.getString("timestamp") ?: at.rocworks.extensions.graphql.TimestampConverter.currentTimeIsoString()
                 )
             } else {
                 // No historical data found, return zero metrics
-                SessionMetrics(0, 0, at.rocworks.extensions.graphql.TimestampConverter.currentTimeIsoString())
+                SessionMetrics(0.0, 0.0, at.rocworks.extensions.graphql.TimestampConverter.currentTimeIsoString())
             }
         })
     }
@@ -326,16 +326,16 @@ class MetricsStoreSQLite(
                 val timestamp = Instant.parse(row.getString("timestamp"))
                 val metricsJson = JsonObject(row.getString("metrics"))
                 val metrics = BrokerMetrics(
-                    messagesIn = metricsJson.getLong("messagesIn", 0L),
-                    messagesOut = metricsJson.getLong("messagesOut", 0L),
+                    messagesIn = metricsJson.getDouble("messagesIn", 0.0),
+                    messagesOut = metricsJson.getDouble("messagesOut", 0.0),
                     nodeSessionCount = metricsJson.getInteger("nodeSessionCount", 0),
                     clusterSessionCount = metricsJson.getInteger("clusterSessionCount", 0),
                     queuedMessagesCount = metricsJson.getLong("queuedMessagesCount", 0L),
                     topicIndexSize = metricsJson.getInteger("topicIndexSize", 0),
                     clientNodeMappingSize = metricsJson.getInteger("clientNodeMappingSize", 0),
                     topicNodeMappingSize = metricsJson.getInteger("topicNodeMappingSize", 0),
-                    messageBusIn = metricsJson.getLong("messageBusIn", 0L),
-                    messageBusOut = metricsJson.getLong("messageBusOut", 0L),
+                    messageBusIn = metricsJson.getDouble("messageBusIn", 0.0),
+                    messageBusOut = metricsJson.getDouble("messageBusOut", 0.0),
                     timestamp = at.rocworks.extensions.graphql.TimestampConverter.instantToIsoString(timestamp)
                 )
                 resultList.add(timestamp to metrics)
@@ -395,8 +395,8 @@ class MetricsStoreSQLite(
                 val timestamp = Instant.parse(row.getString("timestamp"))
                 val metricsJson = JsonObject(row.getString("metrics"))
                 val metrics = SessionMetrics(
-                    messagesIn = metricsJson.getLong("messagesIn", 0L),
-                    messagesOut = metricsJson.getLong("messagesOut", 0L),
+                    messagesIn = metricsJson.getDouble("messagesIn", 0.0),
+                    messagesOut = metricsJson.getDouble("messagesOut", 0.0),
                     timestamp = at.rocworks.extensions.graphql.TimestampConverter.instantToIsoString(timestamp)
                 )
                 resultList.add(timestamp to metrics)
